Split CMS to private trait interface, icc_profile
The purpose of the trait is to be careful with the boundary to the
`moxcms` dependency. We use it because it is a quality implementation
but it is heavy weight for what we need. There's other possible ways to
provide transfer functions and color space transforms. Now this also
introduces ICC profile parsing but again that could be done with a
*much* lighter dependency as we only need basic information from it. The
trait should make every little additional cross-dependency a conscious
decision.

Also it should be the start of a customization point, by feature flag or
actually at runtime.
